backend/requirements.txt
fastapi==0.100.0
psycopg2-binary==2.9.9
pydantic==2.5.2
starlette==0.27.0
uvicorn==0.28.1
peewee==3.17.1
unidecode==1.3.8
python-dotenv==1.0.1
faker==25.2.0